requirements.txt
gymnasium
numpy
torch
matplotlib
pytest
python-dotenv
openai
streamlit
fastapi
uvicorn